refactor: renamed front api objects to be more specific
