+++ web/src/api/index.js
@@ -1,11 +1,13 @@
-import suggestions from './suggestions/suggestions';
-import meetings from './meetings/meetings';
-import tags from './tags/tags';
-import events from './events/events';
+import suggestion from './suggestion/suggestion';
+import meeting from './meeting/meeting';
+import tag from './tag/tag';
+import event from './event/event';
+import user from './user/user';
 
 export default {
-  suggestions,
-  meetings,
-  tags,
-  events
+  suggestion,
+  meeting,
+  tag,
+  event,
+  user
 };
