+++ web/src/api/user/user.js
@@ -1,0 +1,5 @@
+import { get } from '../utils';
+
+export default {
+  getUserId: () => get({ resource: '/' })
+};
